sshfs-map: fix segfaults under musl.
- line and line_size should be reset each loop, otherwise getline() will
use a freed memory address. This worked under glibc, but broke under
musl for some reason. More specifically, it would segfault when running
setmntent for the second time.

- General clean-up: is_mounted checks shouldn't call continue, otherwise
they will skip the clean-up routines. This would be a memory leak, but
not a functionality issue. The calls to continue were probably the
reason that the program didn't segfault when running for a consecutive
time, after the first map had already been mounted.

- Move printf calls to outside fork block.